Struct substrate_state_machine::ProvingBackend

pub struct ProvingBackend<'a, S: 'a + TrieBackendStorage<H>, H: 'a + Hasher> { /* fields omitted */ }

Patricia trie-based backend which also tracks all touched storage trie values. These can be sent to remote node and used as a proof of execution.

Methods

impl<'a, S: 'a + TrieBackendStorage<H>, H: 'a + Hasher> ProvingBackend<'a, S, H>

pub fn new(backend: &'a TrieBackend<S, H>) -> Self

Create new proving backend.

pub fn new_with_recorder(
backend: &'a TrieBackend<S, H>,
proof_recorder: Rc<RefCell<Recorder<H::Out>>>
) -> Self

Create new proving backend with the given recorder.

pub fn extract_proof(self) -> Vec<Vec<u8>>

Consume the backend, extracting the gathered proof in lexicographical order by value.